requirements.txt
psycopg2-binary==2.8.4
psycopg2==2.8.4
passlib==1.7.2
uvicorn==0.11.2
SQLAlchemy==1.3.12
fastapi==0.45.0
pydantic==1.3
starlette==0.12.9
PyJWT==1.7.1
bcrypt==3.1.7
python-multipart==0.0.5
requests-toolbelt==0.7.0
requests==2.22.0
pytest==5.3.2